feat(#24): add Falcosidekick integration
this commit introduces a new field to the Options, `FalcosidekickURL`,
which specifies the URL of the Falcosidekick endpoint to which teler-waf's
events will be forwarded. If this field is set, detected threats will be
logged to teler-waf's logs and forwarded to Falcosidekick for further processing.

Additionally, this commit adds a new method to the Teler, `sendLogs`, which
sends the logs of detected threats to the Falcosidekick instance specified
in `FalcosidekickURL`.

These changes enable users to integrate teler-waf with their existing
Falcosidekick setup, improving their security posture by centralizing threat
detection and response.