fix: nil pointer deref `sendFalcoEvents` func
The previous code in the `sendFalcoEvents()` function attempted to
defer the closing of the response body w/o checking if the HTTP request
had resulted in an error. This could lead to a "nil pointer dereference"
error when an error occurred during the request.

This commit fixes the issue by checking if an error occurred and deferring
the response body close only if there was no error.

Signed-off-by: Dwi Siswanto <me@dw1.io>